Move filesystem function to grb_fs
Reduces tools mixture
