+++ CMakeLists.txt
@@ -194,6 +194,7 @@ add_library(libgerbera STATIC
         src/util/executor.h
         src/util/generic_task.cc
         src/util/generic_task.h
+        src/util/grb_fs.cc
         src/util/grb_fs.h
         src/util/jpeg_resolution.cc
         src/util/logger.h
